requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
airportsdata==20250224
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.9.0
appdirs==1.4.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
beartype==0.20.2
bitsandbytes==0.45.5
blessed==1.20.0
blis==1.3.0
blobfile==3.0.0
brotli==1.0.9
bypy==1.8.9
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
cmake==4.0.0
compressed-tensors==0.9.4
confection==0.1.5
contourpy==1.3.2
cuda-bindings==12.9.0
cuda-python==12.9.0
cycler==0.12.1
cymem==2.0.11
datasets==3.5.0
decorator==5.2.1
decord==0.6.0
deepspeed==0.16.6
diffusers==0.33.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
filelock==3.17.0
flashinfer-python==0.2.5
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
gitdb==4.0.12
gitpython==3.1.44
gmpy2==2.2.1
google-auth==2.40.1
google-genai==1.15.0
gpustat==1.1.1
h11==0.14.0
hf-transfer==0.1.9
hjson==3.1.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.7
importlib-metadata==8.6.1
importlib-resources==6.5.2
inflect==7.5.0
interegular==0.3.3
ipdb==0.13.13
ipython==8.35.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema-specifications==2025.4.1
jsonschema==4.23.0
kiwisolver==1.4.8
langcodes==3.5.0
language-data==1.3.0
lark==1.2.2
lightning-utilities==0.14.3
lit==18.1.8
litellm==1.70.2
llguidance==0.7.21
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
markdown==3.8
markupsafe==3.0.2
matplotlib-inline==0.1.7
matplotlib==3.10.1
mdurl==0.1.2
mergedeep==1.3.4
modelscope==1.26.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
murmurhash==1.0.12
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numpy==2.0.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
open-clip-torch==2.32.0
openai==1.75.0
opencv-python==4.11.0.86
orjson==3.10.18
outlines-core==0.1.26
outlines==0.1.11
packaging==25.0
pandas==2.2.3
parso==0.8.4
partial-json-parser==0.2.1.1.post5
pexpect==4.9.0
pillow==11.1.0
pip==25.0
preshed==3.0.9
prometheus-client==0.22.0
prompt-toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
pyarrow-hotfix==0.6
pyarrow==19.0.1
pyasn1-modules==0.4.2
pyasn1==0.6.1
pycocoevalcap==1.2
pycountry==24.6.1
pycparser==2.22
pycryptodomex==3.23.0
pydantic-core==2.33.1
pydantic==2.11.3
pygments==2.19.1
pynvml==12.0.0
pyparsing==3.2.3
pyre-extensions==0.0.32
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests-toolbelt==1.0.0
requests==2.32.3
rich==14.0.0
rpds-py==0.25.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
sentry-sdk==2.26.1
setproctitle==1.3.5
setuptools==75.8.0
sgl-kernel==0.1.2.post1
sglang==0.4.6.post4
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy==3.8.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.1
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.1
triton==3.2.0
typeguard==4.4.2
typer==0.15.3
typing-extensions==4.12.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.3.0
uv==0.7.6
uvicorn==0.34.2
uvloop==0.21.0
wandb==0.19.9
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
xgrammar==0.1.19
xxhash==3.5.0
yarl==1.19.0
zipp==3.21.0
